Moved receipt related classed to the top package and renamed the clases
